clean and refine kernels
